requirements.txt
pandas==*
numpy<2
newsapi-python==*
requests==*
transformers==*
torch==*
groq==*
langchain-groq==*
langchain==*
langchain-core==*
streamlit==*
tensorflow==*